+++ go.mod
@@ -1,6 +1,6 @@
-module lexido
+module github.com/micr0-dev/lexido
 
-go 1.22.0
+go 1.22
 
 require (
 	github.com/charmbracelet/bubbles v0.18.0
